Consolidated jump scripts #24
