+++ flappy-bird/Assets/Jump.cs
@@ -6,15 +6,20 @@ public class Jump : MonoBehaviour
     public Rigidbody2D Rb;
     public Lifetime Lifetime;
 
+    public int jumpForce = 375;
     public float FallMultiplier = 2.5f;
     public float LowJumpMultiplier = 2f;
 
     InputAction JumpAction;
 
+    private bool canJump;
+
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
         JumpAction = InputSystem.actions.FindAction("Jump");
+        JumpAction.Enable();
+        canJump = true;
     }
 
     // Update is called once per frame
@@ -22,14 +27,24 @@ public class Jump : MonoBehaviour
     {
         if (Lifetime.IsRunning)
         {
-            if (Rb.linearVelocity.y < 0)
-            {
-                Rb.linearVelocity += (FallMultiplier - 1) * Physics2D.gravity.y * Vector2.up * Time.deltaTime;
-            }
-            else if (Rb.linearVelocity.y > 0 && !JumpAction.IsPressed())
-            {
-                Rb.linearVelocity += (LowJumpMultiplier - 1) * Physics2D.gravity.y * Vector2.up * Time.deltaTime;
-            }
+            if (Rb.linearVelocity.y < 0) { Rb.linearVelocity += (FallMultiplier - 1) * Physics2D.gravity.y * Time.deltaTime * Vector2.up; }
+            else if (Rb.linearVelocity.y > 0 && !JumpAction.IsPressed()) { Rb.linearVelocity += (LowJumpMultiplier - 1) * Physics2D.gravity.y * Time.deltaTime * Vector2.up; }
+
+            if (JumpAction.IsPressed()) OnJumpPress();
+            else OnJumpRelease();
         }
     }
+
+    public void OnJumpPress()
+    {
+        Rb.gravityScale = 1;
+        if (canJump)
+        {
+            Rb.linearVelocityY = 0f;
+            Rb.AddForce(new() { x = 0, y = jumpForce });
+        }
+        canJump = false;
+    }
+
+    public void OnJumpRelease() { canJump = true; }
 }
