--- flappy-bird/Assets/JumpInput.cs
@@ -1,45 +1,0 @@
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class JumpInput : MonoBehaviour
-{
-    public Rigidbody2D rb;
-    public int jumpForce;
-
-    public Lifetime Lifetime;
-
-    InputAction jumpAction;
-
-    bool canJump;
-
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        jumpAction = InputSystem.actions.FindAction("Jump");
-        jumpAction.Enable();
-        canJump = true;
-    }
-
-    // Update is called once per frame
-    void Update() 
-    {
-        if (Lifetime.IsRunning)
-        {
-            if (jumpAction.IsPressed()) OnJumpPress();
-            else OnJumpRelease();
-        }
-    }
-
-    public void OnJumpPress() 
-    {
-        rb.gravityScale = 1;
-        if (canJump) 
-        {
-            rb.linearVelocityY = 0f;
-            rb.AddForce(new() { x = 0, y = jumpForce }); 
-        }
-        canJump = false;
-    }
-
-    public void OnJumpRelease() { canJump = true; }
-}
