--- flappy-bird/Assets/JumpInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c454f35f47e8e357fb2977a292754365
