+++ flappy-bird/Assets/Scenes/SampleScene.unity
@@ -536,7 +536,6 @@ GameObject:
   - component: {fileID: 1438963038}
   - component: {fileID: 1438963039}
   - component: {fileID: 1438963041}
-  - component: {fileID: 1438963042}
   - component: {fileID: 1438963043}
   m_Layer: 0
   m_Name: PlayerObject
@@ -706,23 +705,9 @@ MonoBehaviour:
   m_EditorClassIdentifier: Assembly-CSharp::Jump
   Rb: {fileID: 1438963038}
   Lifetime: {fileID: 1295850062}
+  jumpForce: 375
   FallMultiplier: 2.5
   LowJumpMultiplier: 2
---- !u!114 &1438963042
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1438963035}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c454f35f47e8e357fb2977a292754365, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::HandleJumpBehavior
-  rb: {fileID: 1438963038}
-  jumpForce: 375
-  Lifetime: {fileID: 1295850062}
 --- !u!114 &1438963043
 MonoBehaviour:
   m_ObjectHideFlags: 0
